+++ tests/upipe_ts_si_generator_test.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2015-2017 OpenHeadend S.A.R.L.
+ * Copyright (C) 2015-2018 OpenHeadend S.A.R.L.
  *
  * Authors: Christophe Massiot
  *
@@ -116,7 +116,8 @@ static char *iconv_wrapper(void *_unused, const char *psz_encoding,
     /* converted strings can be up to six times larger */
     i_out_length = i_length * 6;
     p = psz_string = malloc(i_out_length);
-    if (iconv(iconv_handle, &p_string, &i_length, &p, &i_out_length) == -1) {
+    if (iconv(iconv_handle, &p_string, &i_length, &p, &i_out_length) ==
+        (size_t)-1) {
         fprintf(stderr, "couldn't convert from %s to %s (%m)\n", psz_encoding,
                 psz_native_encoding);
         free(psz_string);
